Refactor to support forking in RPC endpoints [WIP]
