+++ crates/starknet-devnet-core/src/starknet/get_class_impls.rs
@@ -2,7 +2,7 @@ use blockifier::state::state_api::StateReader;
 use starknet_rs_core::types::BlockId;
 use starknet_types::contract_address::ContractAddress;
 use starknet_types::contract_class::ContractClass;
-use starknet_types::felt::{ClassHash, Felt};
+use starknet_types::felt::ClassHash;
 
 use crate::error::{DevnetResult, Error, StateError};
 use crate::starknet::Starknet;
@@ -13,15 +13,15 @@ pub fn get_class_hash_at_impl(
     block_id: &BlockId,
     contract_address: ContractAddress,
 ) -> DevnetResult<ClassHash> {
-    let state = starknet.get_mut_state_at(block_id)?;
-    let class_hash = state.get_class_hash_at(contract_address.try_into()?)?;
+    let state = starknet.get_mut_state_at(block_id)?; // TODO what if block in origin
+    let core_address = contract_address.try_into()?;
 
-    let class_hash_felt = class_hash.into();
-    if class_hash_felt == Felt::default() {
-        return Err(Error::ContractNotFound);
-    }
+    let class_hash = match state.get_class_hash_at(core_address)? {
+        class_hash if class_hash != Default::default() => class_hash,
+        _ => starknet.defaulter.get_class_hash_at(core_address)?,
+    };
 
-    Ok(class_hash_felt)
+    Ok(class_hash.into())
 }
 
 pub fn get_class_impl(
